+++ test/root_test.dart
@@ -9,7 +9,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:provider/provider.dart';
 
-import 'package:todotada/themenotifier.dart';
+import 'package:todotada/themenotifier/themenotifier.dart';
 import 'package:todotada/main.dart';
 
 ///
